+++ python/__init__.py
@@ -47,6 +47,7 @@ from .check_address import check_address
 from .check_ao40_uncoded_crc import check_ao40_uncoded_crc
 from .check_astrocast_crc import check_astrocast_crc
 from .check_cc11xx_crc import check_cc11xx_crc
+from .check_crc16_ccitt import check_crc16_ccitt
 from .check_crc import check_crc
 from .check_eseo_crc import check_eseo_crc
 from .check_swiatowid_crc import check_swiatowid_crc